+++ HISTORY.rst
@@ -6,7 +6,8 @@ History
 ------------------
 * Add new checks for metrics calculations
 * Add the Sequential Predictive Conformal Inference (SPCI) method for MapieTimeSeriesRegressor.
-
+* Add the Coverage Width-based Criterion (CWC) metric.
+* Add new checks for metrics calculations
 
 0.7.0 (2023-09-14)
 ------------------